EBookBackendSqliteDB: Added e_book_backend_new_contact[s]() APIs
The new API allows control on whether or not to overwrite existing contacts
with the same UID.

Deprecated the older e_book_backend_add_contact[s]() apis which include
the unused 'partial_contact' parameter and have no control on whether
or not to overwrite existing contacts.

Additionally this patch exposes the E_BOOK_SDB_ERROR domain and a couple
of error codes to distinguish if an error occurred because of a constraint
or other errors (UID conflicts at contact addition time are constraint
errors).