Bug #691126 - Fix warnings found by Clang (POP3 part)
The change itself fixed a symbol clash on CAMEL_POP3_STREAM_LINE,
where one was an enum value, and the other a line size integer.
It may or may not change behaviour of the POP3 provider.
This is committed separately, and only to master branch, to give
the change more time for testing.