Make ECollectionBackend and ECollectionBackendFactory instantiable.
ECollectionBackendFactory will use a backend name of "none" and will
instantiate an ECollectionBackend, which takes no action on its own.

Believe it or not, there's actually a use case for this.

GOA will soon add a provider for generic IMAP/SMTP accounts.  Our GOA
module will support these accounts by configuring a collection source
with a backend name of "none" as well as mail account, mail identity,
and mail transport sources.

The mechanics of ESourceRegistryServer do require a collection backend
to support a collection source but the backend will have nothing to do.