ECalCLient test-client-get-free-busy: Giving up on the second round.
This test requires that a custom fixture create a mail account and identity
source in order for e_cal_backend_mail_account_is_valid() to pass
in the backend when it handles the get_free_busy() call.

Somehow we are not able to achieve this easily with
e_source_registry_commit_source_sync(), even added timeouts dont work
for the second round.