ECalClient: Follow up on a timezone-related FIXME comment.
Emit the "timezone-added" signal from an idle callback on ECalClient's
GMainContext, instead of from whichever thread the timezone was added.
This makes life easier for any handlers of that signal.

Still need to address a similar FIXME comment in ECalBackend, but that
will have to wait until after the backend API rewrite.