ECalClient: Use g_bus_watch_name_on_connection().
g_bus_watch_name_on_connection() handles "NameOwnerChanged" signals
as well as "closed" signals from the GDBusConnection.  In the event
the bus name vanishes, we schedule an idle callback on the client's
GMainContext to emit a "backend-died" signal.

Also while we're at it, call the close() method asynchronously from
dispose() so we don't block.