Splits libebook into two libraries.
This clears the way for Direct Read Access APIs to be integrated,
since EBookClient now needs to access EDataBook apis, a common
library is needed to share all the addressbook related objects
related to EContact so that it can be used by both the server
side addressbook backends and EDataBook, as well as EBookClient.

Essentially this properly splits things up and avoids a nasty
circular dependency issue.