Added concurrent view test: test-client-view-operations.c
This tests concurrent access to EBookClientView, 20 threads each
race to receive notifications from the same addressbook and the
test asserts that all 5 contacts in the test addressbook are found.

This is particularly interesting with DEBUG_DIRECT set (Direct Read Access),
as it ensures that concurrent direct reads of the addressbook do not
cause any lockups.