Store E.164 phone number values automatically in the vCard
The E.164 normalized phone number is of interest for handset related
applications. With this change a X-EVOLUTION-E164 attribute is added
to each TEL attribute if the contact summary contains a E.164 formatted
variant of the phone number. This shall avoid overhead and inconsistency
that would occur if clients would use their own mechanism to compute
that already stored information.

Updated Mathias Hasselmann's patch on bug 689622 to apply to current master.