EBackend: Add a "connectable" property (GSocketConnectable).
This is the socket endpoint for the network service to which the
EBackend is a client.  This can be NULL if the EBackend does not
use network sockets.

The initial value of the "connectable" property is derived from the
ESourceAuthentication extension of the EBackend's "source" property,
if the extension is present.