Ported various EBookClient tests to additionally test Direct Read Access variants
This patch basically adds additional invocations of the same test code
in direct read access mode. Only for tests concerned with read access
to the addressbook.

Note that the test-client-uid-only-view and test-client-revision-view
tests are not ported, they are not expected to pass in Direct Read Access
mode (since the EBookClientView in Direct Read Access mode fetches contact
data directly from the backend, any contact field filtering would be a
useless overhead).