Brute force backport of EBookBackendSqliteDB + phone number work.
This is basically overwriting our openismus-work phone-number related
work (plus the EBookBackendSqliteDB) from master, ensuring that the
phone number details work the same way in master and in openismus-work.

As for the direct port of EBookBackendSqliteDB, additional work was done
there, it's more stable, mutexes work properly and I trust it more
(plus it conforms to the same API, and has phone number related work
from Mathias).