e_source_registry_server_load_directory: Monitor failure is non-fatal.
Directory monitoring is a nice-to-have feature, but is not supported in
some contexts.  If we fail to create a GFileMonitor, leave a breadcrumb
on the console to indicate something whent wrong, but don't return an
error status.  That would cause the whole registry process to terminate.

(cherry picked from commit af31c05e026bf9b7881ddc9adfa29d78e29766c0)