Adapted EDataBook error conversions from master into openismus-work branch
For some reason, the D-Bus error conversions are working fine when
the client is connected with default access. When the client is connected
in DRA mode and has it's own copy of EDataBook in memory, the error codes
sent over D-Bus dont wrap/unwrap correctly... this fixes bug 694188 for
DRA mode in the openismus-work branch.