Add e_source_registry_dup_unique_display_name().
Compares a source's display name against other sources having an
extension with the given extension name, or else against all other
sources in the registry if no extension name was given.

If the source's display name is unique among these other sources, the
function will return the source's display name verbatim.  Otherwise the
function will construct a string that includes the source's own display
name as well as those of its ancestors.

The functions's return value is intended to be used in messages shown to
the user to help clarify which source is being referred to.  It assumes
the source's display name is at least unique among its siblings.