build: Support (and prefer) libboost_thread-mt
Google's libphonenumber uses Boost.Thread but somehow fails to let their
shared library to express that dependency. Therefore we must manually
link with Boost.Thread. Now Fedora only installs libboost_thread-mt, but
not libboost_thread. This patch checks for both build variants of the
library and prefers libboost_thread-mt. The change in preference is
because it doesn't make much sense to use a Boost.Thread variant that
isn't thread-safe, does it?