build: Support (and prefer) libboost_thread-mt
Google's libphonenumber uses the Boost Thread library but somehow fails
to let its shared library express that dependency. Therefore we must
manually link that library. Now Fedora only installs libboost_thread-mt,
but not libboost_thread. This patch checks for both build variants of
the threads library and prefers libboost_thread-mt.

The change in preference is because it doesn't make much sense to use
a thread support library that isn't thread-safe, does it?

https://bugzilla.gnome.org/show_bug.cgi?id=694749