EBookBackendSqliteDB: Add new overall revision to the contacts DB
Previous upgrades overlooked the fact that we normalized the contact
data in the summary columns, this added revision ensures that all
contact data is normalized (something that was already done for a
change in phone number normalizations, but only for version 4 -> 5).

For some added safety, this patch changes the normalization
routine to be nested into a transaction.

Fixes bug 695145