EDataBookFactory: Rewrite client connection tracking.
Keep a hash table of client bus names mapped to a set of EBookBackend
references.  When the backend emits a closed() signal with the client's
bus name, we remove that EBookBackend reference from the data structure.

Also keep a hash table of client bus names we're watching so that if a
bus name vanishes we can remove all EBookBackend references associated
with it.