EDataCalFactory: Rewrite client connection tracking.
Keep a hash table of client bus names mapped to a set of ECalBackend
references.  When the backend emits a closed() signal with the client's
bus name, we remove that ECalBackend reference from the data structure.

Also keep a hash table of client bus names we're watching so that if a
bus name vanishes we can remove all ECalBackend references associated
with it.