Create only one EDataBook per EBookBackend.
I've never understood why we create multiple EDataBook instances per
EBookBackend.  All it's doing is exporting an AddressBook interface on
the session bus and handling incoming method invocations by forwarding
them to the EBookBackend.

The GDBusMethodInvocation identifies the client, so we don't need to
export a separate AddressBook interface per client.  This seems to be
a relic of the Bonobo-era design.

Fixing this redundancy is a prerequisite to running backends in child
processes, which is one of my goals for 3.10.

New functions:

    e_book_backend_ref_data_book()
    e_book_backend_set_data_book()

Deprecated functions:

    e_book_backend_add_client()
    e_book_backend_remove_client()