Create only one EDataCal per ECalBackend.
I've never understood why we create multiple EDataCal instances per
ECalBackend.  All it's doing is exporting a Calendar interface on the
session bus and handling incoming method invocations by forwarding them
to the ECalBackend.

The GDBusMethodInvocation identifies the client, so we don't need to
export a separate Calendar interface per client.  This seems to be a
relic of the Bonobo-era design.

Fixing this redundancy is a prerequisite to running backends in child
processes, which is one of my goals for 3.10.

New functions:

    e_cal_backend_ref_data_cal()
    e_cal_backend_set_data_cal()

Deprecated functions:

   e_cal_backend_add_client()
   e_cal_backend_remove_client()