Let EDataBook track the "opened" state.
Due to the asymmetric nature of our backend APIs, the open() method
begins in EBookBackend, but the end result is reported to EDataBook.
So it makes more sense for EDataBook to track the backend's "opened"
state, at least until the APIs can be modernized.

This adds a new function -- e_data_book_is_opened() -- which is just a
temporary hack and should only be called by e_book_backend_is_opened().

This avoids EDataBook needing to call e_book_backend_notify_opened(),
which is deprecated and about to be removed.