Let EDataCal track the "opened" state.
Due to the asymmetric nature of our backend APIs, the open() method
begins in ECalBackend, but the end result is reported to EDataCal.
So it makes more sense for EDataCal to track the backend's "opened"
state, at least until the APIs can be modernized.

This adds a new function -- e_data_cal_is_opened() -- which is just a
temporary hack and should only be called by e_cal_backend_is_opened().

This avoids EDataCal needing to call e_cal_backend_notify_opened(),
which is deprecated and about to be removed.