Revert "EGdbusTemplates: Address crash on operation cancel"
This reverts commit 7c52fbd81093c1264e3d8aa6cdcf5c8bdc7b1772.

7c52fb introduced a memory leak because it indirectly creates a new
GQuark for each synchronous method call. The string for that GQuark
will never get freed. In addition, the table for quarks keeps growing
and also leaks. This is what valgrind found, the strings themselves
are still reachable and therefore not reported by valgrind.

The bigger problem is that with 7c52fb, e_client_open fails with
EDB_OPENING_ERROR = "Cannot process, book backend is opening" in the
"SYNCEVOLUTION_NO_PIM_EDS_DIRECT=1 testpim.py
TestContacts.testFilterStartup" test. When using DRA, the problem is
that the client seems to use the wrong DB path.

This can be either a race condition that is merely triggered by the
change in 7c52fb or a real problem in the patch itself. Either way,
until a solution is found it is better to not use the patch in the
openismus-work branch.

Conflicts:
	libedataserver/e-gdbus-templates.c

The conflict was due to g_idle_add <-> g_idle_add_full. Resolved by using
g_idle_add_full with NULL pointers.