data_book_convert_to_client_error: add E_DATA_BOOK_STATUS_BUSY
This maps to E_CLIENT_ERROR_BUSY.

This error was encountered in evolution-addressbook-factory
when multiple clients open the same address book at the same
time. The missing error mapping was not causing problems,
because D-Bus error mapping elsewhere handled it and ensured
that the EBookClient got the right code. The only downside
was triggering the "code should not be reached" warning in the
default case.

Adding the error is more important for DRA. However, in that
case it is uncertain whether this error can occur - better add it
anyway.