EBookBackendSqliteDB: Avoid summary introspection on initial creation.
The summary introspection does not introspect whether regular prefix
indexes are created, which is acceptable since it does not effect queries.

However since we introspect before creating the indexes, we lose the
configured index information which is needed to initially create those
indexes.

This patch simply checks if the table exists when initially creating
the summary and avoids the introspection if the table is being created
for the first time.