Added test case test-sqlite-cursor-set-target.c
This test case tests e_book_backend_sqlitedb_cursor_set_target().

The special case of setting a partial state is tested so that when
a cursor is setup to sort by "family_name, given_name" and the target
is set to "J", then we recieve results inclusive of the contact with
the family name "J".