EBookBackendSqliteDB: Added localized data to all string columns in the summary.
This updates the schema to version 7, adds columns in the contacts tables to
hold collated data (using g_utf8_collate_key()) and adds an lc_collate column
to the folders table to track the currently effective LC_COLLATE locale name.

The database is recollated at any time that the LC_COLLATE domain change
is detected.