Improve error handling in imapx_server_get_message().
If a GET_MESSAGE job for the same UID is already in progress, wait for
the job to finish, but don't assume it was successful.  Try retrieving
the message from cache, but if that fails then we have to re-fetch the
message from the IMAP server.

(cherry picked from commit cb048105f08625cf504bd8a539cafeccc92ba89e)