Bug 699223 - Emit view signals from context where view was started
This applies to both EBookClientView and ECalClientView.

Incoming D-Bus signals were being propagated to the EClient's main loop
context, but there may not be a running main loop on that thread, so the
main loop sources were never being dispatched and the view signals never
emitted.

Instead use the main loop context which was the thread-default at the
time the view was started, which is often the so-called "main" thread
where a main loop is more likely to be running.

(cherry picked from commit 417164c733128072ea3e4ad685829f0c3270c1b7)

Conflicts:
	addressbook/libebook/e-book-client-view.c
	calendar/libecal/e-cal-client-view.c