ECalBackendWeather: e_weather_source_new() is failable.
Backend logic was failing to take this into account and crashing if the
weather location code in the ESource was invalid.

I'm not sure if EWeatherSource is even worth keeping around, but I'll
come back to that later.  Not doing an all-out rewrite at the moment.

(cherry picked from commit ef17c11c438c390378130ae942d81d6feb65405d)