Mega-Commit changing the API of fetching the initial addressbook locale.
This is a bit of a messy patch to support the same API backported into 3.6,
what we've done is extend the addressboook 'open' D-Bus API to report the
locale of the addressbook initially. Using the previously created get_locale()
D-Bus API was broken as it caused lockups when an addressbook is opened from
a dedicated thread (as test-client-view-operations does).

This patch is just a huge workaround to avoid the lockup in
test-client-view-operations.