Remove CamelIMAPXConnManager.
It turns out CamelIMAPXConnManager's server selection logic has been
broken for some time and we've been operating on a single connection
regardless of the "concurrent-connections" setting.

And you know what?  A single connection works fine.  The only thing
CamelIMAPXConnManager actually does nowadays is occasionally deadlock
because it tries to connect to a server while holding a locked mutex,
which is a major no-no.

CamelIMAPXConnManager is not documented, it's not only NOT doing its
job but is actively causing problems, and I can't seem to wrap my head
around the implications of multiple connections in a CamelService.  So
I decided to just junk it.

New functions:

  camel_imapx_store_ref_server()

Removed functions:

  camel_imapx_store_get_server()
  camel_imapx_store_op_done()