CamelTcpStreamRaw: Don't timeout on reads.
This is needed for IMAP IDLE to work, which sits in a read call for
long periods of time.  If this times out then we disconnect from the
server and miss out on mailbox updates.

The timeout was orignally added to work around suspend/resume issues,
but that should be handled by way of GCancellables, not by prematurely
terminating a read call.  Prior to reading we connect to GCancellable
and call PR_Interrupt() on cancellation to terminate the read call.
Need to make sure the GCancellable is actually firing when it should,
and that we're not racing it.

Leaving the timeout on write calls in place for now, since that seems
more reasonable.