[NNTP] Fails to get list of folders, if not connected already
The code, incorrectly, assumed that the store is always connected
when a list of folders is requested, but the actual connection
happens only during command invocation, after whose successful
finish it's ensured that there is defined also a stream.

This can be reproduced by creating a new NNTP account and trying
to manage subscription. The first try fails, console is filled with
runtime warnings, the second try shows list of available folders.