Bug 705206 - Stale "Retrieving message xxx" in preview panel
A client application cancelling an operation should not terminate the
parser thread.  The parser thread completes the operation even if the
client is no longer interested in the result.  The parser thread loop
must only be broken out of on an I/O error, parsing error, the server
hanging up, or application shutdown.