[Maildir] Create the 'tmp' directory, if missing
If a folder's directory is missing 'tmp' sub-directory, but it has
all the other 'cur' and 'new' directories, then try to create
the 'tmp' to get all three required directories.

It's because some backup tools can drop the 'tmp' directories.

This was reported as: https://bugzilla.redhat.com/show_bug.cgi?id=995702