Return a GError in camel_imapx_job_wait().
Despite the improvements in commit dacf161, there is still a disconnect
in error propagation between CamelIMAPXCommand and CamelIMAPXJob in that
completed jobs are still not aware of whether they failed.

This commit lays the foundations for another attempt to bridge the gap,
this time by adding a means of setting an error on a CamelIMAPXJob with
camel_imapx_job_take_error().  We also break API in order to return the
error through camel_imapx_job_wait().

New functions:

  camel_imapx_job_take_error()