imapx_command_start() should not be failable.
This might by why error handling in CamelIMAPXServer is so confusing.

The imapx_command_start() function itself should not be failable.  If an
error occurs while starting the given command, stop the parser thread so
it disconnects from the IMAP server, and embed the GError in the command
structure with camel_imapx_command_failed().

I suspect this will simplify the IMAPX dispatching logic significantly.