imapx_command_start_next() should not be failable.
If we fail to stop an in-progress IDLE command, the parser thread and
all unfinished commands will terminate on their own.  Don't propagate
back a GError for this particular failure, because the caller is just
trying to enqueue a new CamelIMAPXCommand and doesn't really know how
to handle the error.