Add camel_imapx_list_response_dup_attributes().
Returns a GHashTable of mailbox attributes to be used as a set.
Use g_hash_table_contains() and g_hash_table_get_keys() to query
for and list all mailbox attributes, respectively.

Also, internally we now normalize added mailbox attributes to match
the CAMEL_IMAPX_LIST_ATTR_* macros and track them with a GHashTable
instead of a GQueue so duplicates are naturally eliminated.