CamelIMAPXStoreSummary: Remove all public members.
The 'version' member was not used for anything and we no longer use the
'namespaces' member.  Only the CamelIMAPXServer knows about server-side
mailbox namespaces.  That information is not relevant when offline, and
therefore doesn't need to be cached.

This also removes imapx_fabricate_old_namespace_list() and similar hacks
in CamelIMAPXServer.