Bug 708730 - CalDAV: Improve handling of invalid access tokens
If e_source_get_oauth2_access_token_sync() fails, rather than dumping
the error message to stderr where users won't see it, stash the GError
within the CalDAV backend to be retrieved when the backend falls back
to caldav_authenticate().

This both returns a descriptive error message to the client application
and prevents an inappropriate password prompt while authenticating with
an access token instead of a password.