Bug 710787 - Replace most uses of sprintf() with g_snprintf()
sprintf() can potentially overwrite the buffer, but g_snprintf() takes
a length to stop it from doing that.

Some of these might instead be replaced by g_strdup_printf().  It is
hard to know if the use of alloca() and sprintf() was just because
g_strdup_printf() did not exist yet, or if alloca() was used for
performance.