Bug 710668: ESource: Verify removal before returning
Similar to bug 685986 when creating ESources, e_source_remove_sync()
now waits for an "object-removed" signal from GDBusObjectManagerClient
before returning.  But because ESource has no direct connection to the
"object-removed" signal, it waits for the ESourceRegistry to strip the
ESource of its D-Bus proxy via __e_source_private_replace_dbus_object().

Here again, to avoid risking a deadlock we place a limit on the wait
time (two seconds).  If this time limit expires, we return from the
function regardless.