Update libebook-contacts::test-untyped-phones test
The test_business has been disabled, because it was failing, but it
was failing even in time of 2.28. Looking briefly into RFC 2426
shows that the 'VOICE' is a default 'TEL' type, but there is no
indication that the left-alone 'WORK' is supposed to be the same
as 'WORK,VOICE' type. Due to this are done these changes.