CamelOfflineStore: Prefer service's "host-reachable" status.
For some stores, the remote host machine may still not be reachable even
if the network is available.  Common example is a host machine requiring
a virtual private network connection to reach.

If the CamelOfflineStore is also a CamelNetworkService, and the service
indicates the remote host is not reachable, then always return FALSE in
camel_offline_store_get_online().

Also emit "notify::online" along with "notify::host-reachable".