Ported the lower level cursor tests to use EBookSqlite.
Refactored the whole test suite so that it's isolated to
libedata-book, use the EBookSqlite directly.

Now for the majority of test cases, we test with and without
fields in the summary (with an empty summary configuration)
and we also test the EBookSqlite with a callback to fetch
vcards and without a callback.

We also now test for e164 number change notifications in locale changes.

Conflicts:

	tests/libebook/data/vcards/sorted-4.vcf
	tests/libebook/data/vcards/sorted-5.vcf
	tests/libedata-book/Makefile.am