Bug 709838 - CamelDataCache: Empty cache file leads to deadlock
camel_data_cache_get() forgot to abort the key reservation from its
object bag when an empty cache file is detected.  This would eventually
lead to a deadlock when attempting to add the same key later.

This bug was mainly affecting HTML mails, when a remote image download
got interrupted and left behind an empty file in the HTTP cache.