test-client-cursor-operations.c: Test cursor in fallback mode
Now test the cursor functionality in cases where the cursor
is using sort fields not in the summary (and search expressions
not in the summary).

Also added --filter option, pass a regular expression to filter
which test cases should run

Also increased TIMEOUT tollerance. This needs fixing, backends are not
shutting down properly, locale changes get backed up and take more and
more time as the test keeps running.