CamelFolder: Remove unused fetch_messages_sync() method.
The method is not used by Evolution and was partially broken anyway.
The boolean return value was overloaded to mean both whether an error
occurred AND whether there were more messages to fetch.  That's more
than two possible states and would have required an additional "out"
parameter to clarify, but I'm not going to bother with it.